+++ modules/esm/jrg-ui/.npmignore
@@ -1,0 +1,4 @@
+src
+node_modules
+build
+index.mjs
