+++ modules/esm/jrg-ui/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@jrg/ui",
+  "version": "0.0.3",
+  "description": "UI Components for my website",
+  "main": "target/jrg-ui.min.esm.mjs",
+  "scripts": {
+    "rollup:watch": "rollup -w -c ./build/rollup.config.mjs",
+    "rollup": "rollup -c ./build/rollup.config.mjs"
+  },
+  "author": "Jackie Gleason",
+  "license": "ISC",
+  "devDependencies": {
+    "@jrg/rollup-sass-shadow": "latest",
+    "@material/mwc-button": "latest",
+    "@material/mwc-drawer": "latest",
+    "@material/mwc-icon": "latest",
+    "@material/mwc-icon-button": "latest",
+    "@material/mwc-list": "latest",
+    "@material/mwc-top-app-bar-fixed": "latest",
+    "@rollup/plugin-commonjs": "latest",
+    "@rollup/plugin-node-resolve": "latest",
+    "@xstate/fsm": "latest",
+    "lit-element": "^2.4.0",
+    "mariadb": "latest",
+    "marked": "latest",
+    "mustache": "latest",
+    "rollup": "latest",
+    "rollup-plugin-html": "latest",
+    "rollup-plugin-string": "latest",
+    "rollup-plugin-terser": "latest",
+    "rxjs": "latest",
+    "tslib": "latest"
+  }
+}
