+++ modules/esm/jrg-ui/src/MarkdownComponent/index.mjs
@@ -69,7 +69,7 @@ export class MarkdownComponent extends ShadowElement{
             this.wrapLinks();
             const sty = document.createElement("link");
             sty.setAttribute("rel", "stylesheet");
-            sty.setAttribute('href',"/assets/github-markdown.css");
+            sty.setAttribute('href',"//unpkg.com/@jrg/github-markdown-css/github-markdown.css");
             const sty2 = document.createElement("style");
             sty2.innerHTML = style;
             this.shadowRoot.innerHTML = null;
