+++ modules/esm/jrg-ui/src/SidebarComponent/index.mjs
