+++ modules/esm/jrg-ui/src/SplashComponent/index.mjs
